Removing Interfaces

+++ BattleBot/Constructs/ClientArena.cs
@@ -4,8 +4,8 @@ namespace BattleBot
 {
     public class ClientArena
     {
-        public double Size { get; set; }
-        public double Capacity { get; set; }
+        public decimal Size { get; set; }
+        public decimal Capacity { get; set; }
         public List<WorldObject> Obstacles { get; set; } = new List<WorldObject>();
         public List<Bot> Bots { get; set; } = new List<Bot>();
         public Bot ClientBot;
